+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Versioning].
 ## [Unreleased]
 
 - (`fced786`) Allow `require()` for `no-top-level-side-effects`.
+- (`c46dec9`) Allow `Symbol()` for `const` in `no-top-level-variables`.
 - (`8e4566b`) Optionally allow top-level `const` assignments of arrays and
   objects.